backend/requirements.txt
asgiref==3.4.1
Django==3.2.5
pytz==2021.1
sqlparse==0.4.1
djangorestframework==3.11.0
django-cors-headers==3.7.0
django-filter~=2.4.0
pillow
djoser~=2.1.0

gunicorn==20.0.4
psycopg2-binary==2.8.6